Menkam

+++ public/js/activite.js
@@ -1,11 +1,15 @@
+var page = 1;
+var current_page = 1;
+var total_page = 0;
+var is_ajax_fire = 0;
+var datacurrent_page;
+var effectif_classe = 0;
+var idActivi = 0;
+
+
+
 $(document).ready(function(){
-    var page = 1;
-    var current_page = 1;
-    var total_page = 0;
-    var is_ajax_fire = 0;
-    var datacurrent_page;
-    var effectif_classe = 0;
-    var idActivi = 0;
+
     //alert(dateCourante+" "+heureCourante);
 
     $.ajaxSetup({
@@ -341,6 +345,7 @@ function getPageDataMatiereActivity(id,typeActivite) {
         var id_enseignant;
 
 
+
         if(typeActivite=="normale" || typeActivite=="rattrapage"){
             dateMatiere = $("#dateMatiereActivite").val();
             id_creneau = $("#id_creneauMatiere").val();
@@ -410,6 +415,7 @@ function getPageDataMatiereActivity(id,typeActivite) {
 
         function addExamen(id_session){
             //alert("url "+url+" id "+id_activite+" dateMatiere "+dateMatiere+" id_creneau "+id_creneau+" id_matiere "+id_matiere+" id_surveillant "+id_surveillant+" id_session "+id_session+" type activite"+typeActivite);
+            
             $.ajax({
                 dataType: 'json',
                 type:'POST',
@@ -497,16 +503,20 @@ function getPageDataMatiereActivity(id,typeActivite) {
     $("body").on("click",".removeActivity",function(){
         var id = $(this).parent("td").data('id');
         var c_obj = $(this).parents("tr");
-
-        $.ajax({
-            dataType: 'json',
-            type:'delete',
-            url: url + '/' + id
-        }).done(function(data){
-            c_obj.remove();
-            toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
-            getPageDataActivity();
-        });
+        if(confirm('Voulez-vous vraiment supprimer cette Activite ?')){
+            $.ajax({
+                dataType: 'json',
+                type:'delete',
+                url: url + '/' + id
+            }).done(function(data){
+                c_obj.remove();
+                toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
+                getPageDataActivity();
+            });
+        }else{
+            tostAvertissement("Oppération Annulée");
+        }
+        
     });
 
     // Remove Matière activite
@@ -516,18 +526,22 @@ function getPageDataMatiereActivity(id,typeActivite) {
         var c_obj = $(this).parents("tr");
 
         //alert("l'ide = "+id);
-        $.ajax({
-            dataType: 'json',
-            type:'delete',
-            url: 'delMatiereExamen',
-            data: {
-                id:id
-            }
-        }).done(function(data){
-            c_obj.remove();
-            toastr.success('Matière Deleted Successfully.', 'Success Alert', {timeOut: 5000});
-            getPageDataMatiereActivity(id_activite);
-        });
+        if(confirm('Voulez-vous vraiment supprimer cette Matiere?')){
+            $.ajax({
+                dataType: 'json',
+                type:'delete',
+                url: 'delMatiereExamen',
+                data: {
+                    id:id
+                }
+            }).done(function(data){
+                c_obj.remove();
+                toastr.success('Matière Deleted Successfully.', 'Success Alert', {timeOut: 5000});
+                getPageDataMatiereActivity(id_activite);
+            });
+        }else{
+            tostAvertissement("Oppération Annulée");
+        }    
     });
 
 
@@ -537,18 +551,22 @@ function getPageDataMatiereActivity(id,typeActivite) {
         var id = $(this).parent("td").data('id');
         var c_obj = $(this).parents("tr");
 
-        $.ajax({
-            dataType: 'json',
-            type:'delete',
-            url: 'delClasseActivite',
-            data: {
-                id:id
-            }
-        }).done(function(data){
-            c_obj.remove();
-            toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
-            getPageClasseActivity(id_activite);
-        });
+        if(confirm('Voulez-vous vraiment supprimer cette Classe ?')){
+            $.ajax({
+                dataType: 'json',
+                type:'delete',
+                url: 'delClasseActivite',
+                data: {
+                    id:id
+                }
+            }).done(function(data){
+                c_obj.remove();
+                toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
+                getPageClasseActivity(id_activite);
+            });
+        }else{
+            tostAvertissement("Oppération Annulée");
+        }
     });
 
     // Remove salle activite
@@ -557,18 +575,22 @@ function getPageDataMatiereActivity(id,typeActivite) {
         var id = $(this).parent("td").data('id');
         var c_obj = $(this).parents("tr");
 
-        $.ajax({
-            dataType: 'json',
-            type:'delete',
-            url: 'delSalleActivite',
-            data: {
-                id:id
-            }
-        }).done(function(data){
-            c_obj.remove();
-            toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
-            getPageDataSalleActivity(id_activite);
-        });
+        if(confirm('Voulez-vous vraiment supprimer ?')){
+            $.ajax({
+                dataType: 'json',
+                type:'delete',
+                url: 'delSalleActivite',
+                data: {
+                    id:id
+                }
+            }).done(function(data){
+                c_obj.remove();
+                toastr.success('Activity Deleted Successfully.', 'Success Alert', {timeOut: 5000});
+                getPageDataSalleActivity(id_activite);
+            });
+        }else{
+            tostAvertissement("Oppération Annulée");
+        }
     });
